third_party/blink/web_tests/external/wpt/service-workers/cache-storage/cache-matchAll.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] Cache.matchAll with responses containing "Vary" header
  assert_equals: Cache.matchAll should match the entire header if a vary header is present in both the query and cached requests. expected 1 but got 0
[FAIL] Cache.matchAll with multiple vary pairs
  assert_equals: Cache.matchAll should support multiple vary request/response pairs. expected 3 but got 1
Harness: the test ran to completion.
